requirements.txt
streamlit==1.24.1
pytesseract==0.3.10
Pillow==10.0.0
fpdf==1.7.2
matplotlib==3.7.1
numpy==1.26.4
pdf2image==1.16.3
opencv-python==4.7.0.72
